Updated app
Added a clock that updates in real time as well as a "Smart Action" feature which allows the user to add/modify/remove calendar events simply by typing in what they wish to do.

This is controlled by via the OpenAI API being able to access a SQLite database of user-defined events (excludes auto-generated holidays). The GPT then generates SQLite commands based on the user's request that are recognized and executed by the app to ultimately perform the action desired by the user